fix: make the Windows build link under MSVC, and close three review findings
The Windows CI jobs finally ran (earlier pushes kept cancelling them via
cancel-in-progress) and found a genuine MSVC-vs-MinGW difference plus two
pre-existing warnings, none of which a MinGW build can surface.

sync_menu is a WINDOWS-subsystem executable so no console ever appears,
but the toolchains disagree about which symbol starts it there: the MSVC
CRT looks for wWinMain and fails to link with an unresolved WinMain when
only main() exists, while MinGW supplies a shim that calls main(). The
body is now an ordinary function with a toolchain-appropriate entry point
in front of it.

The warnings-as-errors job was also failing on my own CI bug: passing
-DCMAKE_CXX_FLAGS replaces the defaults CMake supplies on Windows, which
include /EHsc, so every file using exceptions emitted C4530 and /WX turned
that into a failure everywhere. /EHsc is now explicit, with a note about
why it cannot be dropped.

Two pre-existing warnings that /W4 surfaces because nothing had ever built
this tree with MSVC: a size_t-to-int narrowing in websocket.cpp, made
explicit since the array is fixed at 20 bytes and cannot truncate; and
getenv in the pairing store's Windows default-path branch, where MSVC's
suggested replacement is an MSVC-only CRT extension that would break the
MinGW build to silence a call that is not actually unsafe.

Also from the independent review:

The pairing prompt could hang shutdown. A window handle only exists once
the CBT hook fires, slightly after show() is entered, and a cancel landing
in that gap closed nothing and never looked again -- leaving the dialog up
until a human clicked it with the destructor's join() blocked behind it.
The watcher now keeps asking until the window appears or show() returns.
The regression test hangs against the unfixed code, so it can actually
detect what it claims to.

Reparse-point refusal now has a test that runs. The existing symlink tests
skip silently without Developer Mode, so on most CI images that guarantee
had no executing coverage at all. A junction needs no privilege; the new
test puts one mid-path -- the case FILE_FLAG_OPEN_REPARSE_POINT does not
cover, since it only guards the final component -- and asserts the
junction genuinely resolves first, so a pass means the store refused it.

CompanionProcess's destructor depends on its dispatcher still being able
to run callbacks, which was only inferable from a comment in app_main.cpp.
That contract is now stated on dispatch_to_owner itself.

+++ native/src/websocket.cpp
@@ -391,7 +391,11 @@ std::string websocket_accept_key(std::string_view key) {
     return {};
   }
   std::array<unsigned char, 29> encoded{};
-  const int encoded_length = EVP_EncodeBlock(encoded.data(), digest.data(), digest.size());
+  // digest is a fixed 20-byte array, so this conversion cannot truncate;
+  // made explicit because EVP_EncodeBlock takes an int and MSVC /W4
+  // correctly refuses to narrow silently.
+  const int encoded_length = EVP_EncodeBlock(encoded.data(), digest.data(),
+                                             static_cast<int>(digest.size()));
   if (encoded_length != 28) {
     return {};
   }
